+++ src/fides/api/schemas/connection_configuration/__init__.py
@@ -92,6 +92,9 @@ from fides.api.schemas.connection_configuration.connection_secrets_saas import (
 from fides.api.schemas.connection_configuration.connection_secrets_saas import (
     SaaSSchemaFactory as SaaSSchemaFactory,
 )
+from fides.api.schemas.connection_configuration.connection_secrets_scylla import (
+    ScyllaSchema, ScyllaDocsSchema,
+)
 from fides.api.schemas.connection_configuration.connection_secrets_snowflake import (
     SnowflakeDocsSchema as SnowflakeDocsSchema,
 )
@@ -130,6 +133,7 @@ secrets_schemas: Dict[str, Any] = {
     ConnectionType.postgres.value: PostgreSQLSchema,
     ConnectionType.redshift.value: RedshiftSchema,
     ConnectionType.saas.value: SaaSSchema,
+    ConnectionType.scylla.value: ScyllaSchema,
     ConnectionType.snowflake.value: SnowflakeSchema,
     ConnectionType.sovrn.value: SovrnSchema,
     ConnectionType.timescale.value: TimescaleSchema,
@@ -179,4 +183,5 @@ connection_secrets_schemas = Union[
     FidesDocsSchema,
     SovrnDocsSchema,
     DynamoDBDocsSchema,
+    ScyllaDocsSchema
 ]